refactor: extract transaction modal controls
